Rename MetadataFileContents to AppMetadataJSON
