--- src/db/models/app/relations/ProjectHasVersions.ts
@@ -1,8 +1,0 @@
-import { VersionRelation } from "../DBVersion";
-import { ProjectSlugRelation } from "../DBProject";
-
-export interface ProjectHasVersions
-  extends VersionRelation,
-    ProjectSlugRelation {
-  id: number;
-}
